add lock checks to remaining endpoints

+++ traffic_ops/traffic_ops_golang/steeringtargets/steeringtargets.go
@@ -204,6 +204,15 @@ func (st *TOSteeringTargetV11) Create() (error, error, int) {
 		return userErr, sysErr, errCode
 	}
 
+	cdn, err := dbhelpers.GetCDNNameFromDSID(st.ReqInfo.Tx.Tx, int(dsID))
+	if err != nil {
+		return nil, errors.New("createSteeringTarget: getting CDN from DS ID " + err.Error()), http.StatusInternalServerError
+	}
+	// CheckIfCurrentUserCanModifyCDN
+	if userErr, sysErr, errCode := dbhelpers.CheckIfCurrentUserCanModifyCDN(st.ReqInfo.Tx.Tx, cdn, st.ReqInfo.User.UserName); userErr != nil || sysErr != nil {
+		return userErr, sysErr, errCode
+	}
+
 	rows, err := st.ReqInfo.Tx.NamedQuery(insertQuery(), st)
 	if err != nil {
 		return api.ParseDBError(err)
@@ -230,6 +239,16 @@ func (st *TOSteeringTargetV11) Update(h http.Header) (error, error, int) {
 	if err != nil {
 		return errors.New("delivery service ID must be an integer"), nil, http.StatusBadRequest
 	}
+
+	cdn, err := dbhelpers.GetCDNNameFromDSID(st.ReqInfo.Tx.Tx, dsIDInt)
+	if err != nil {
+		return nil, errors.New("updateSteeringTarget: getting CDN from DS ID " + err.Error()), http.StatusInternalServerError
+	}
+	// CheckIfCurrentUserCanModifyCDN
+	if userErr, sysErr, errCode := dbhelpers.CheckIfCurrentUserCanModifyCDN(st.ReqInfo.Tx.Tx, cdn, st.ReqInfo.User.UserName); userErr != nil || sysErr != nil {
+		return userErr, sysErr, errCode
+	}
+
 	dsID := uint64(dsIDInt)
 	// TODO determine if the CRUDer automatically does this
 	st.DeliveryServiceID = &dsID
@@ -303,6 +322,14 @@ func (st *TOSteeringTargetV11) Delete() (error, error, int) {
 		return userErr, sysErr, errCode
 	}
 
+	cdn, err := dbhelpers.GetCDNNameFromDSID(st.ReqInfo.Tx.Tx, int(*st.DeliveryServiceID))
+	if err != nil {
+		return nil, errors.New("deleteSteeringTarget: getting CDN from DS ID " + err.Error()), http.StatusInternalServerError
+	}
+	// CheckIfCurrentUserCanModifyCDN
+	if userErr, sysErr, errCode := dbhelpers.CheckIfCurrentUserCanModifyCDN(st.ReqInfo.Tx.Tx, cdn, st.ReqInfo.User.UserName); userErr != nil || sysErr != nil {
+		return userErr, sysErr, errCode
+	}
 	result, err := st.ReqInfo.Tx.NamedExec(deleteQuery(), st)
 	if err != nil {
 		return nil, errors.New("steering target delete exec: " + err.Error()), http.StatusInternalServerError
